Fixed bugs, added disconnection handling, and created complex full round test

+++ server/server-proj/src/server/createServer.js
@@ -5,6 +5,8 @@ import { Server } from "socket.io";
 
 import LobbyRepository from "../repositories/lobby.repository.js";
 import LobbyController from "../controllers/lobby.controller.js";
+import PlayerActionController from "../controllers/playerAction.controller.js";
+import GameEngineService from "../services/gameEngine.service.js";
 
 export function createServer({ corsOrigin = "*" } = {}) {
     const app = express();
@@ -17,6 +19,9 @@ export function createServer({ corsOrigin = "*" } = {}) {
     // Dependencies
     const lobbyRepository = new LobbyRepository();
     const lobbyController = new LobbyController(lobbyRepository);
+
+    const playerActionController = new PlayerActionController(lobbyRepository);
+    let gameEngineService;
 
     const LOBBY_ROOM = "lobby";
 
@@ -31,6 +36,10 @@ export function createServer({ corsOrigin = "*" } = {}) {
     }
 
     io.on("connection", (socket) => {
+        
+        
+        // Lobby events
+        
         socket.on("lobby:join", ({ username } = {}, ack) => {
             try {
                 const { playerId, isHost, lobby } = lobbyController.lobbyJoin(username);
@@ -74,6 +83,11 @@ export function createServer({ corsOrigin = "*" } = {}) {
 
                 // TODO: Start game engine here, then emit initial game state:
                 // io.to(LOBBY_ROOM).emit("game:state", initialState);
+
+                gameEngineService = new GameEngineService(lobbyRepository.players); 
+
+
+
             } catch (err) {
                 ack?.({ ok: false, error: err.message });
             }
@@ -91,23 +105,53 @@ export function createServer({ corsOrigin = "*" } = {}) {
             );
         });
 
+
+        // Game events
+
+        socket.on("game:action", ({ playerId, action, amount } = {}, ack) => {
+            try {
+                playerActionController.performAction(playerId, action, amount);
+
+
+                ack?.({ ok: true });
+            } catch (err) {
+                ack?.({ ok: false, error: err.message });
+            }
+        });
+
+
         socket.on("disconnect", () => {
             try {
-                const { playerId, username } = socket.data || {};
+                
+                if (lobbyRepository.isGameStarted) {
+                   
+                    
+                    // TODO: Handle mid-game disconnects 
+                    
+                        // Check if only one player left - if so, end the game and reset lobby
+                        // If game is ongoing, mark player as disconnected and skip their turns until they reconnect or game ends
+                        // If player was host, assign new host (could be first player to the left in seating order, or first to join lobby, etc)
+                        
+                    
 
-                // If this socket never joined the lobby, ignore it
-                if (!playerId) {
-                    return;
                 }
+                else {
+                    const { playerId, username } = socket.data || {};
 
-                if (playerId && !lobbyRepository.isGameStarted) {
-                    lobbyRepository.removePlayer(playerId);
-                    emitLobbyState();
+                    // If this socket never joined the lobby, ignore it
+                    if (!playerId) {
+                        return;
+                    }
+
+                    if (playerId && !lobbyRepository.isGameStarted) {
+                        lobbyRepository.removePlayer(playerId);
+                        emitLobbyState();
+                    }
+
+                    console.log(
+                        `Player ${username} (${playerId}) disconnected and removed from lobby`
+                    );
                 }
-
-                console.log(
-                    `Player ${username} (${playerId}) disconnected and removed from lobby`
-                );
             } catch (err) {
                 console.error("Error removing player on disconnect:", err);
             }
